+++ src/components/Calculator/elements/CalculatorHeader/CalculatorHeader.tsx
@@ -1,6 +1,7 @@
 import { GenerationNum, Result } from '@smogon/calc';
+import type { TFunction } from 'i18next';
 import React, { useCallback, useState } from 'react';
-import { TFunction, useTranslation } from 'react-i18next';
+import { useTranslation } from 'react-i18next';
 import Button from 'semantic-ui-react/dist/commonjs/elements/Button';
 import Icon from 'semantic-ui-react/dist/commonjs/elements/Icon';
 import Dropdown from 'semantic-ui-react/dist/commonjs/modules/Dropdown';